move enum Seat to utils/enums.ts
