+++ src/utils/enums.ts
@@ -1,0 +1,5 @@
+export enum Seat {
+	AVAILABLE,
+	BOOKED,
+	CHOOSED,
+}
